requirements.txt
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
openai>=1.0.0,<2.0.0
python-dotenv>=0.19.0,<2.0.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.5,<1.0.0
httpx>=0.24.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.10.0,<2.0.0 